// src/lib/types.ts
export enum CommentRefType {
  Post = 0,
  Event = 1,
  News = 2,
  Moment = 3,
}

export enum InteractionType {
  Sad = 0,
  Happy = 1,
  Love = 2,
  Angry = 3,
  Wow = 4,
  Like = 5,
  DisLike = 6,
}
export enum InteractionRefType {
  Story = 0,
  News = 1,
  Moment = 2,
}